move bigint serialization from common to serialization package
